requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-cors-headers==3.6.0
django-rest-framework==0.1.0
djangorestframework==3.12.1
idna==2.10
pycodestyle==2.6.0
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
toml==0.10.1
urllib3==1.25.10
